+++ changelog.md
@@ -5,6 +5,7 @@
 * Added Rainbow Wool Stairs & Slab
 * Prismarine blocks can now be used as in conduit structure
 * Food crates and bread boxes can no longer be composted
+* Paper and Dried Kelp can no longer be used as fuel (vanilla behavior that was modified by the mod)
 * Added `beveled_glass_panes` block tag
 * Added `patterned_wool_stairs` & `patterned_wool_slabs` block/item tags
 * Removed now unused `nether_portal_frame_blocks` block tag